Correct if statement so it returns true
The if statement to check for an existing artist would only return
false and never true. This meant that an album could not be saved even
with a valid artist id.

Fix this by adding an else branch that returns true if artist find by id
is not nil (false).